[BEAM-8376] Google Cloud Firestore Connector - Add common classes
In preparation, this change adds common classes used by each of the three
sub-components of the Firestore Connector. The three sub-components to follow
are:
1. Firestore v1 Read operations
2. Firestore v1 Write operations
3. Firestore SDK Adapter

### General Common classes
* FirestoreDnFn: base DoFn for all Firestore DoFn
* FirestoreIO: DSL Entry point
* FirestoreIOOptions: Those options which are configurable at the PipelineOptions level
* FirestoreStatefulComponentFactory: Factory for those components used by DoFn which are stateful and provides a mockable boundary in unit tests
* JodaClock: Simplistic clock interface which provides a mockable boundary

### Integration Tests

A GCP Service account with read & write permissions for Cloud Firestore is
required and the `GOOGLE_APPLICATION_CREDENTIALS` environment variable should be
configured to point to the service account file. Alternatively, the integration
test suite can be ran against the Cloud Firestore emulator by setting
`FIRESTORE_EMULATOR_HOST` to the host-port of the running Cloud Firestore
Emulator, and setting `GOOGLE_CLOUD_PROJECT` to the name of the project that
will be used by the emulator.

All integration tests are integrated with JUnit tests lifecycle and will have
any documents or collections created during the course of a test run cleaned up
after the test finishes running (regardless of success). See the
`org.apache.beam.sdk.io.gcp.firestore.it.FirestoreTestingHelper` JUnit rule for
more details.