[BEAM-8376] Google Cloud Firestore Connector - Add RPC QoS Classes
In preparation, this change adds RPC QoS classes used by the Firestore v1 Read
and Write operations.

### Rpc Quality of Service (QoS)
In an effort to not overwhelm the live Cloud Firestore service, client side
RPC Quality of Service has been built into FirestoreV1. All requests: attempted,
successful and failed are recorded and used to determine the rate at which
requests are sent to Cloud Firestore. In the case of `Write` requests, writes
are enqueued and flushed once a batch size threshold is reached, or a bundle
finishes. A combination of Adaptive Throttling[1], 500/50/5 Ramp Up[2] and
historical batch throughput are used to determine the maximum size of a batch.

Per RPC Method metrics will be surfaced allowing for insight into the values
that feed into the QoS System.

For each RPC Method the following counters are available:
* throttlingMs
* rpcFailures
* rpcSuccesses
* rpcStreamValueReceived

[1] https://sre.google/sre-book/handling-overload/#client-side-throttling-a7sYUg
[2] https://cloud.google.com/firestore/docs/best-practices#ramping_up_traffic