[BEAM-12060] Fix failing Go Postcommits, jenkins support for Gradle tasks.
This fixes the Flink and Spark Postcommits failing after some gradle tasks were changed. This was due to not enabling Jenkins support for the changed gradle tasks for the Postcommits which run on Jenkins. In addition to fixing that, this change adjusts the way that Jenkins support is enabled. Instead of creating a whole new task to run integration tests with Jenkins, the tasks now check a project property.