Warn when using a deterministic fallback coder
Deterministic coders are required by some transforms, such as for the
key part of an element in a GroupByKey or CombinePerKey.

The DeterministicFastPrimitivesCoder is adequate for most types, but if
you get this warning it means that there might be a faster type-specific
coder. Usually setting the proper type hint helps.

https://beam.apache.org/documentation/sdks/python-type-safety/#use-of-type-hints-in-coders