[BEAM-12009] Implement Calc splitting rule.
This required a few modifications from Calcite:
- "The latest level should be CalcRelType otherwise literals cannot be
implemented"
https://github.com/apache/calcite/blob/8581f0a3fe9a4f079cb4d36f02121ae22118714c/core/src/main/java/org/apache/calcite/rel/rules/CalcRelSplitter.java#L395-L404
This restriction seems to be overly specific to Calcite's
internal usage of CalcRelSplitter. Either Beam RelType can implement
literals, so it doesn't matter what the latest level is.
- "Sometimes a level's program merely projects its inputs. We don't want these.
They cause an explosion in the search space."
https://github.com/apache/calcite/blob/8581f0a3fe9a4f079cb4d36f02121ae22118714c/core/src/main/java/org/apache/calcite/rel/rules/CalcRelSplitter.java#L232-L237
This check causes failure for queries that don't read from a table. We
create a dummy input row in Beam. This results in a BeamZetaSqlCalcRel
being created that only reads that dummy input row. This check detects
that the program is "trivial" and discards it. However, the next level
then fails an assertion because it does not reference the dummy row.
https://github.com/apache/calcite/blob/d9a81b88ad561e7e4cedae93e805e0d7a53a7f1a/core/src/main/java/org/apache/calcite/rex/RexProgram.java#L466
So basically this is a bug in the check. I'm not sure if there's a way
to improve it so it works as intended, but removing it doesn't seem to
hurt us any (not sure why this would cause an "explosion in the search
space"; in our case, it just results in an extra ZetaSQL rel, which the
planner cannot further optimize anyway).
- Level count must be at least 1, otherwise certain programs, such as
those that perform only a project, will not have any Calc implementation
created and thus will fail.

[BEAM-12009] Fix bugs in calc splitting implementation.