[BEAM-12204] Allow unboundedly large side inputs in portable Java. (#14604)
We now cache just the (decoded) first page of elements read over
the state API, rather than the whole thing.

There is a slight change of semantics here: the pages must be delimited
at element boundaries.  This is already the case for elements sent over
the data channel, other langauges, and all runners, but required some
changes to tests.