Remove blank line that javadoc doesn't like
The blank line in the example `private static class MyDataSourceProviderFn` in the "Reading from JDBC datasource" causes javadoc to terminate prematurely. While this is probably a bug in javadoc, it causes the documentation to be missing everything after that spot -- including the rather important "how to write to a JDBC datasource" section.. If you take out this line the javadoc builds correctly. This bug was not in 2.23.0 but was in 2.24.0