Warn on BigQuery Job failures
Currently if a BigQuery job were to fail because of a data issue, the failures are logged at INFO level. This makes it challenging for users to debug because through the runner (e.g. Dataflow) everything appears fine. Logging at WARN or ERROR will give users more indication that something is wrong with the pipeline.