[BEAM-9920][BEAM-11418] Go Xlang artifact staging to GCP w/ Dataflow (#14748)
Adjust the Go Dataflow runner to stage cross-language artifacts to GCP as well, and adds them to the list of packages for workers. This fixes a few failing cross-language pipelines that required artifacts to be read. Also refactors the artifact staging code slightly to allow for staging remotely in a more generic fashion, and not just to GCP.