[BEAM-8888] Improve handling of unknown types in CalciteUtils.
Also enables the nullness checker to make sure we didn't miss anything.

Because of the design of the Row/Schemas API, we unfortunately have to
add a bunch of checks. These fields are probably not allowed to be null
during Schema construction, but the checker framework does not know
that.