[BEAM-12402] Optimize PCollectionConsumerRegistry$MultiplexingMetricTrackingFnDataReceiver by making immutable list.
In a Nexmark benchmark profile 2% of cpu was spent on
AbstractMapBasedMultimap$WrappedCollection$WrappedIterator
methods beneath this class.

The get method returns a wrapper around the set that allows for set modification.
This is more expensive to iterate and is also undesired because we don't want
to allow the list to be modified.

We iterate over this list many times for counters. It appears that we don't need a wrapped list as it is documented that all consumers should be registered first. So it seems we can just create a copy there to an immutable list and trivially save that cpu.