feat: read change stream partition dofn (#18)
* refactor: centralises initial partition checks

Creates a InitialPartition class to do any verifications related to the
initial partition.

* fix: allow nulls in mods for data records

Old values and new values can be null in a data record.

* fix: allow nulls as values in mods

Old values and new values in mods can contain null values that need to
be properly serialized with Avro. We added a schema definition that
allows for such case here.

* fix: fix partition tracker trySplit

* feat: adds stop mode for partition position

* feat: remove initial partition logic from tracker

Removes the special casing for the initial partition in the restriction
tracker. We only allow splits when there is an claim for a timestamp
that is at least 1 second in the future.

* feat: add timestamp converter class

* feat: extract partition restriction splitter class

This will be responsible for encapsulating the logic to trying to split
a certain restriction.

* chore: add reminder to add tests

* feat: extract restriction claimer class

* test: add test for timestamp converter class

* test: add test to restriction claimer

We also start checking for preconditions here.

* test: add test for restriction splitter

* feat: add split checker for restriction

* test: add test for restriction tracker

* fix: fix partition restriction constructors

Fixes the wait for child partitions restriction constructor to send the
children to wait for and adds tests.

* feat: add logging to the restriction split checker

* feat: add logging to the restriction splitter

* fix: fix restriction claim logic

Fixes the restriction claimer to consider the last claimed position in
addition to the restriction. Fixes the restriction claimer to always
return false when trying to claim from a STOP mode.

* fix: increase default heartbeat interval

The default of 1 second was returning an error.

* fix: allow claiming end of interval

* test: fix spanner change stream it test

* feat: changes the log level of change streams

To info level for now

* chore: formats the code