fix: fix split for wait for child / parents (#22)
When we are splitting a wait for child partitions or a wait for parent
partitions we need to make sure the restriction residual does not skip a
phase, when scheduled before the primary. This is problematic in the
case where the primary resumes, such as in the wait for child partitions
/ wait for parent partitions.
In this PR, when splitting one of the two aforementioned modes, we
remain at the same mode, instead of advancing. This will make sure the
residual only proceeds if the current phase is complete.