fix: fix child record processing (#28)
We used to wait for children to be scheduled on every child partition
record received. The problem with that is that a child was not
scheduled, it could result in resuming the function with the mode
WAIT_FOR_CHILD_PARTITION. When resumed, the function would just wait for
the children and start finishing the partition. Because of this, any
records after the first child partition record for the same stream would
be in fact ignored.

This PR changes the logic, so that the waiting for child partitions to
be scheduled is only done when the change stream ends (meaning all child
partition records have been processed).