[BEAM-12740] [BEAM-8268] Improve error handling and retries for GCS rename operations
Previously exceptions such as timeouts or unavailable for a single file would not be caught
and would cause retries and backoff of the entire DoFn processing.

Additionally plumb the options to ignore missing sources through to underlying FileSystem.rename
so that file systems can handle such errors intelligently per-file.  The filtering in FileSystems.rename
to remove such files is beneficial because it may use cheaper match operations but it is insufficient
as previous deletes observed as timeouts may delete files between the match and rename.

Further reduce operations by modifying FileBasedSink to not delete files that have been renamed.