[BEAM-5379] Fix Go SDK proto generation, and update protos.
Proto generation was broken from the Go module suffix change to /v2. This fixes that, as well as updates the version of protoc-gen-go used, updates proto generation instructions to be accurate with modules, and regenerates all the protos. Note that this latest version of protoc-gen-go seperates GRPC generated code into separate files with the suffix _grpc.pb.go.