fix: wait for children to be running / finished  (#62)
* fix: fix split checker

The previous logic was comparing the last claimed position timestamp
against the timestamp of the current claimed position. The problem with
this is that if the write rate is high, Spanner will have multiple
updates at the same timestamp. Because of this, the split checker will,
more often than not, return false, preventing splitting from happening.

In this PR we compare the claimed position timestamp against the
restriction start timestamp. In this sense, we allow for splits if at
least 1 microsecond has passed from the start timestamp. This fixes the
issue of having multiple records with the same timestamp.

* fix: use watermark estimator in bundle finalizer

Instead of registering the callback with the current watermark, gives
the watermark estimator to the bundle finalizer callback so that the
current watermark is retrieved within the callback.

* feat: add heartbeat record count metric

Adds optional metric to check for heartbeat count.

* feat: add debug logging to restriction tracker

* fix: wait for children to be running / finished

Instead of waiting for children to be scheduled / running / finished,
wait for children to be either running or finished. This is because when
they are scheduled they do not hold the watermark yet.