Avoid setting empty builders in proto setters
This is a minor upstream change from a Google internal cleanup. The idea is to avoid setting empty builders in proto setters, when setting the default instance would do just as well. This makes the code cleaner, and, at scale, more efficient.

In this particular case, it’s probably what you wanted to begin with, as the 4 functions surrounding this one already use this pattern