[BEAM-12383] Adding Go SDK and Kafka IO to Gradle cross-language test suites. (#15094)
* [BEAM-12383] Adding the Go SDK to Gradle cross-language test suites.

With this change, Go SDK cross-language tests should automatically run on the the Cross-Language PostCommit test suites on Jenkins. There are also some changes to the Kafka IO test in particular to fix bugs and get it to work better, mostly centered around how it starts up and shuts down the Kafka cluster.