Allow multiple Python worker processe to share the same VM.
The use of this feature is guarded by a new SDK capability.

This is an alternative to starting up multiple docker images to get
around the fact that the GIL often prevents more than one CPU to be
saturated by a worker.  It has the benefit of amortizing the overhead
of managing and setting up VMs across multiple worker processes.

The notion of "siblings" is introduced, because the container needs
to be started up with an explicit worker id, which needs to be
provided to identify itself to the runner for retrieval of provisioning
information.  The runner can then respond with any number of additional,
worker ids that indicate identical workers should be started up with
this same provisioning information.