fix: avoid using read as transaction can get stuck (#77)
* fix: avoid using read as transaction can get stuck

There is a bug in the current java-spanner client library used, where
if specific conditions are met, a transaction might get stuck while
performing a read call. This has been fixed in later versions of the
client library, but Apache Beam still uses a version without the fix.

In order to work around the issue, we do not use any read calls, but
instead do the same with a streaming execute query instead.

* feat: add todo to rollback to read in dao

Adds todo to rollback to use read when java-spanner library is updated
to contain the fix.