Allow wildcards for java class lookup transform providers.
This is especially useful for cases where the expansion service is implicitly
started by the user for ad-hoc invocation of cross-language transforms (in which
case there is no security benefit to an explicit allow list).

Also automatically starts up an expansion service for Python's
JavaExternalTransform and adds the ability to manually specify a classpath.