[BEAM-13164] Address most of a race condition between instantiation and usage of member variables.
There is a race condition being observed where we are passing the `InboundObserver` to the `outboundObserverFactory` here (https://github.com/apache/beam/blob/bee60cdc29d995170d4461692e391683ca7dcafd/sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/data/BeamFnDataGrpcMultiplexer2.java#L68). This allows for incoming calls to start before the BeamFnDataGrpcMultiplexer2 is fully initialized meaning that the erroredInstructionIds is accessed before it is initialized.

This seems like a general problem with how we connect the inbound and the outbound observer (because of the ClientCallStreamObserver callack that is necessary) that only happens on the SDK harness since there is no need to connect the inbound and outbound observers this way on the ServerCallStreamObserver.