Make S3 streaming more efficient.
Modifies the Boto3 client to keep track of a previously opened HTTP
stream instead of creating a new connection/range request for each read.

This implementation is more cache efficient also for the server, since
it allocates only one long-lived stream for the whole S3 object instead
of requesting a new range with each read.