Use shorter, numerical indices for dataframe reads.
We attempt to match user expectations by giving every element across all
shards a unique index. Previously this was done by embedding the filepath
itself in the index, but this would result in this (often quite long) path
being duplicated for every element (sometimes exceeding the size of the
data itself).

As we expand the full set of shards before doing the actual reading,
we can use an upper bound on the count to shift the indices of each shard
to ensure uniqueness. A nice, round power of ten is used for the shift to
visually make it easy to see where the indices come from if they are ever
exposed to the user.

In addition, fixes a bug where the file splitting was not taken into
account when doing index uniquification.