[BEAM-13052] Add Pub/Sub Lite xlang transforms in python (#15727)
* [BEAM-13052] Restructure pubsublite folder to move non-user interface classes to internal folder

PubsubLiteIO is composed of quite a few classes that are not part of the user interface for this module. Adding support for x-lang transforms will make this worse.

Move everything that is not part of the intended user interface to the `internal` folder.

* [BEAM-13052] Restructure pubsublite folder to move non-user interface classes to internal folder

PubsubLiteIO is composed of quite a few classes that are not part of the user interface for this module. Adding support for x-lang transforms will make this worse.

Move everything that is not part of the intended user interface to the `internal` folder.

* [BEAM-13052] Add Pub/Sub Lite xlang transforms

* [BEAM-13052] Add Pub/Sub Lite xlang transforms

* [BEAM-13052] Add Pub/Sub Lite xlang transforms

* Revert unexpected merge conflict

* Reformat

* fix java checkstyle

* fix license headers

* fix formatting

* fix lint

* fix merge

* [BEAM-13052] Restructure pubsublite folder to move non-user interface classes to internal folder

PubsubLiteIO is composed of quite a few classes that are not part of the user interface for this module. Adding support for x-lang transforms will make this worse.

Move everything that is not part of the intended user interface to the `internal` folder.

* [BEAM-13052] Restructure pubsublite folder to move non-user interface classes to internal folder

PubsubLiteIO is composed of quite a few classes that are not part of the user interface for this module. Adding support for x-lang transforms will make this worse.

Move everything that is not part of the intended user interface to the `internal` folder.

* Revert unexpected merge conflict