[BEAM-13430] Re-enable dependency analysis for modules.
The upgrade to gradle 7.x caused the migration from 'provided' scope to 'compileOnly' which required more dependencies to be flagged as 'permitUnusedDeclared' as they aren't all necessary for compilation but are necessary for pom.xml generation.
This change also include the migration from hamcrest_core and hamcrest_library to hamcrest since the single package now supersedes it which the dependency analsyis plugin picked up on.
Finally, I also removed a usage of guava within the Dataflow worker code and now depend only on the shaded version of guava within that codebase.