feat: use outputWithTimestamp instead of output (#98)
When producing records from a DoFn uses outputWithTimestamp in order to
correctly assign the record timestamp. We were using only the output
method before and the record timestamp was somehow inherited from the
watermark. The Dataflow team feels this could be a bug in the Beam code,
so we are using the explicit outputWithTimestamp instead.