chore: remove obsolete fixme / todo (#111)
* chore: remove obsolete FIXMEs

The FIXMEs removed were related on what to do if exceptions were thrown.
In the case of a streaming pipeline these will be retried indefinitely.

* chore: remove obsolete TODO

The TODO was related to what to do if an exception was thrown by the
change stream mapper. In this case, the exception will bubble up to the
worker and it will be retried indefinitely. This will allow customers to
realize that something is wrong.