[BEAM-12164]: Add retry protection to DetectNewPartitions
The original algorithm of the DetectNewPartitions is susceptible to
failures, because it produces side effects on every try which is not
idempotent. Specifically, it marks the partitions as SCHEDULED in the
Spanner database and outputs them. If there is a bundle commit failure,
during retry, the already SCHEDULED partitions will not be picked up
again.

We change the algorithm in this PR to always schedule partitions that
have a created at timestamp greater than the one saved in the
DetectNewPartitions restriction. When scheduling the partitions, this
SDF will also claim the created at of such partitions, advancing the
timestamp saved. If there is a bundle commit failure, the restriction
timestamp won't be saved, thus the partitions in the bundle will be
picked up again regardless of their state.