Merge pull request #16655 from [BEAM-12164]: Add retry protection to DetectNewPartitions
* feat: add experimental spanner readChangeStreams

Adds the SpannerIO.readChangeStreams feature that will enable users to
consume a change stream from Cloud Spanner.
This feature is under preview now, and can only be used for allowlisted
customers.

When reading a change stream the users will be able to operate on a
PCollection of DataChangeRecords, containing the modifications made to
the database as well as the type of operation.

* fix: remove public api exposure of opencensus

Do not expose the Opencensus TraceSampler in the
SpannerIO.readChangeStreams.

This is done so that we can upgrade the opencensus library without
having to concern ourselves with which version customers are using.

This commit also removes the deserializer option since it is not used.

* [BEAM-12164]: Add retry protection to DetectNewPartitions

The original algorithm of the DetectNewPartitions is susceptible to
failures, because it produces side effects on every try which is not
idempotent. Specifically, it marks the partitions as SCHEDULED in the
Spanner database and outputs them. If there is a bundle commit failure,
during retry, the already SCHEDULED partitions will not be picked up
again.

We change the algorithm in this PR to always schedule partitions that
have a created at timestamp greater than the one saved in the
DetectNewPartitions restriction. When scheduling the partitions, this
SDF will also claim the created at of such partitions, advancing the
timestamp saved. If there is a bundle commit failure, the restriction
timestamp won't be saved, thus the partitions in the bundle will be
picked up again regardless of their state.

Co-authored-by: Zoe Cai <zoc@google.com>