Avoid packaging avro in the java harness jar.
Unfortunately we can't simply shade it, as it's used in the public API
(AvroCoder) of sdks:java:core.  However, outside that use, it should not
be needed when running in portable mode.  Users will have to provide
their own avro (typically by depending on sdks:java:core, but possibly
choosing another version explicitly or from elsewhere).