BEAM-13159 Update embedded-redis dependency
The io.oden/embedded-redis artifact was provided by me as part of
landing BEAM-13159, as the previously used version did not use a recent
enough version of Redis to support the XADD/XRANGE command set and also
had a fatal crash bug on darwin/arm64 ("Apple Silicon").

My fork in turn was based on the signal.org artifact, which had a more
recent redis version but was still affected by the crash bug.  Now that
https://github.com/signalapp/embedded-redis/pull/12 has landed, the
io.oden artifact is unlikely to be upgraded further; beam should use the
dependency that is actively maintained.