[BEAM-12164]: Fixes SpannerChangeStreamIT
The GSON object was being initialized in a static context. When the DoFn
that used it was called in another JVM, which did not initialized this
object, we were getting a NPE exception.

Here, we move the initialization of the GSON object inside the DoFn
itself (through a setup method).