[BEAM-13106] Prevent infinite wait in Flink savepoint test.
This test starts a job, cancels it, and then restores it from a savepoint. Flink 1.14 revealed a bug in the test. If the initial, canceled job remains listed first in the cluster, the test would keep getting it blindly, and waiting because it was not in state RUNNING (instead it was CANCELED).

I fixed the bug by renaming the pipeline to disambiguate the two jobs.