[BEAM-12164]: allow for nanosecond precision
Nanosecond precision is used by Spanner to track change streams, for
which we are adding support here. Previously we only supported
microsecond precision.

Implements a timestamp restriction tracker on top of bytekeyrange
tracker to allow for desired precision and to claim timestamps. On top
of this, we built two decorator restriction trackers for each of our
SDFs, which behave slightly differently.

The DetectNewPartitionsTracker allows for claiming timestamps equal to
or greater than the current restriction timestamp.
The ReadChangeStreamPartitionTracker does the same as the above, but it
disallows for splitting when the initial (fake) partition is being
queried. We do not want to split at that case, because there is no way
to resume from where we left of specifically for the initial change
stream query.

This commit also modifies the relevant tests.