Merge pull request #16844 from [BEAM-12164]: allow for nanosecond precision
* [BEAM-12164]: allow for nanosecond precision

Nanosecond precision is used by Spanner to track change streams, for
which we are adding support here. Previously we only supported
microsecond precision.

Implements a timestamp restriction tracker on top of bytekeyrange
tracker to allow for desired precision and to claim timestamps. On top
of this, we built two decorator restriction trackers for each of our
SDFs, which behave slightly differently.

The DetectNewPartitionsTracker allows for claiming timestamps equal to
or greater than the current restriction timestamp.
The ReadChangeStreamPartitionTracker does the same as the above, but it
disallows for splitting when the initial (fake) partition is being
queried. We do not want to split at that case, because there is no way
to resume from where we left of specifically for the initial change
stream query.

This commit also modifies the relevant tests.

* fix: do not increment if timestamp is max

We convert the closed-closed interval of the change stream query from
the user to a closed-open interval for the restriction.
If the end of the interval was max and added a nanosecond it would
overflow. This commit prevents the overflow.

* chore: fix underflow bug for previous timestamp

* fix: fix partition finishing

Partitions were not being finished, because the end of the restriction
could not be claimed in the case where the end was MAX timestamp.

We update the end at parameter to be at most MAX timestamp - 1ns, which
prevents this issue from happening.

* fix: fix linting violations

* chore: add quickcheck generators dependency

In order to use primitive generators for property testing

* feat: remove bytekey from timestamp tracker

The behaviour of the bytekeyrangetracker when checkpointing was
generating a duplicate timestamp between the primary and the residual.
This caused duplicate data change records to be emitted.

In this commit, we implement our own tracker from scratch with the
desired behavior.

* chore: minor fixes

* fix: reports timestamp progress in fraction

Reporting the number of nanos left was causing an overflow in windmill.
We normalize the progress here to be between the interval of [0, 1].

* test: add javadocs / tests for timestamp utils

* test: additional test for range tracker / docs

* refactor: refactor / add comments

* chore: fix linting violations

* test: fix partition metadata mapper test

EndAt is no longer allowed to be null (defaults to Timestamp.MAX - 1ns).

* fix: allows for quickcheck packages publicly

Quickcheck is only used in testing. Allow for the classes to be exposed.

* fix: allows for timestamp max as input