Merge pull request #17023 from [BEAM-12164]: Remove child partition query workaround
* [BEAM-12164]: Remove child partition query workaround

The backend used to throw an error if we queried at the exact end
timestamp of a partition (when child records were received). Because of
this, we implemented a workaround that would always query 1us (or 1ns)
before the end time, so that we could get records back.

This bug has been fixed, so here we remove the workaround entirely.

* fix: fix the end timestamp of the partition query

* test: fix spannerio change streams unit test

* fix: fix underflow / overflow on timestamps

Makes sure that timestamp operations take care of underflow and
overflow when initializing restrictions.