[BEAM-14157] Don't call requestObserver.onNext on a closed windmill streams
requestObserver.onNext() which is called by send should not be called after requestObserver.onCompleted() is called.
requestObserver.onCompleted() is called when the stream is closed.

I am investigating stalled windmill streams with logs like `Output channel stalled for 31s, outbound thread`.
I think sending messages on a closed stream has something to do with it.
In test runs with this change, I'm yet to see a stalled GetWorkStream.